requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file dev.txt dev.in
#
click==6.6                # via pip-tools
dj-database-url==0.4.1
django-cors-headers==1.2.1
django-debug-toolbar==1.5
django-extensions==1.7.4
django-filter==0.15.2
Django==1.10.2            # via django-debug-toolbar
djangorestframework-jwt==1.8.0
djangorestframework==3.4.7
first==2.0.1              # via pip-tools
pip-tools==1.7.0
psycopg2==2.6.2
PyJWT==1.4.2              # via djangorestframework-jwt
requests==2.11.1
six==1.10.0               # via django-extensions, pip-tools
spotipy==2.3.8
sqlparse==0.2.1           # via django-debug-toolbar
